Refactor & start writing to /data and use queue worker to print file names
